+++ pvcompare/demand.py
@@ -52,6 +52,8 @@ def calculate_load_profiles(
     static_inputs_directory=None,
     user_inputs_pvcompare_directory=None,
     user_inputs_mvs_directory=None,
+    add_electricity_demand=None,
+    add_heat_demand=None,
 ):
     r"""
     Calculates electricity and heat load profiles for `country`, `storeys`, and `year`.
@@ -91,6 +93,13 @@ def calculate_load_profiles(
         Path to MVS specific input directory. If None,
         `constants.DEFAULT_USER_INPUTS_MVS_DIRECTORY` is used.
         Default: None.
+    add_electricity_demand: str or None
+        Path to precalculated hourly electricity demand time series for one year (or the same period
+        of a precalculated PV timeseries). Default: None.
+    add_heat_demand: str or None
+        Path to precalculated hourly heat demand time series for one year (or the same period
+        of a precalculated PV timeseries). Default: None.
+
 
     Returns
     ------
@@ -115,28 +124,87 @@ def calculate_load_profiles(
     for column in energyConsumption:
         if column != "unit":
             if energyConsumption.at["energyVector", column] == "Heat":
-                calculate_heat_demand(
-                    country=country,
-                    lat=lat,
-                    lon=lon,
-                    storeys=storeys,
-                    year=year,
-                    weather=weather,
-                    static_inputs_directory=static_inputs_directory,
-                    user_inputs_pvcompare_directory=user_inputs_pvcompare_directory,
-                    user_inputs_mvs_directory=user_inputs_mvs_directory,
-                    column=column,
-                )
+
+                if add_heat_demand is not None:
+                    # check if file exists
+                    if os.path.isfile(add_heat_demand):
+                        # save the file name of the time series
+                        # mvs_inputs/elements/csv/energyProduction.csv
+                        check_inputs.add_file_name_to_energy_consumption_file(
+                            column=column,
+                            ts_filename=add_heat_demand,
+                            user_inputs_mvs_directory=user_inputs_mvs_directory,
+                        )
+                    else:
+                        logging.warning(
+                            "The heat demand time series"
+                            " you have entered does not exist. It is"
+                            " thus calculated according to the default"
+                            " method of pvcompare."
+                        )
+                        calculate_heat_demand(
+                            country=country,
+                            lat=lat,
+                            lon=lon,
+                            storeys=storeys,
+                            year=year,
+                            weather=weather,
+                            static_inputs_directory=static_inputs_directory,
+                            user_inputs_pvcompare_directory=user_inputs_pvcompare_directory,
+                            user_inputs_mvs_directory=user_inputs_mvs_directory,
+                            column=column,
+                        )
+                else:
+                    calculate_heat_demand(
+                        country=country,
+                        lat=lat,
+                        lon=lon,
+                        storeys=storeys,
+                        year=year,
+                        weather=weather,
+                        static_inputs_directory=static_inputs_directory,
+                        user_inputs_pvcompare_directory=user_inputs_pvcompare_directory,
+                        user_inputs_mvs_directory=user_inputs_mvs_directory,
+                        column=column,
+                    )
             elif energyConsumption.at["energyVector", column] == "Electricity":
-                calculate_power_demand(
-                    country=country,
-                    storeys=storeys,
-                    year=year,
-                    static_inputs_directory=static_inputs_directory,
-                    user_inputs_pvcompare_directory=user_inputs_pvcompare_directory,
-                    user_inputs_mvs_directory=user_inputs_mvs_directory,
-                    column=column,
-                )
+
+                if add_electricity_demand is not None:
+
+                    if os.path.isfile(add_electricity_demand):
+                        # save the file name of the time series
+                        # mvs_inputs/elements/csv/energyProduction.csv
+                        check_inputs.add_file_name_to_energy_consumption_file(
+                            column=column,
+                            ts_filename=add_electricity_demand,
+                            user_inputs_mvs_directory=user_inputs_mvs_directory,
+                        )
+                    else:
+                        logging.warning(
+                            "The electricity demand time series"
+                            " you have entered does not exist. It is"
+                            " thus calculated according to the default"
+                            " method of pvcompare."
+                        )
+                        calculate_power_demand(
+                            country=country,
+                            storeys=storeys,
+                            year=year,
+                            static_inputs_directory=static_inputs_directory,
+                            user_inputs_pvcompare_directory=user_inputs_pvcompare_directory,
+                            user_inputs_mvs_directory=user_inputs_mvs_directory,
+                            column=column,
+                        )
+                else:
+                    calculate_power_demand(
+                        country=country,
+                        storeys=storeys,
+                        year=year,
+                        static_inputs_directory=static_inputs_directory,
+                        user_inputs_pvcompare_directory=user_inputs_pvcompare_directory,
+                        user_inputs_mvs_directory=user_inputs_mvs_directory,
+                        column=column,
+                    )
             else:
                 logging.warning(
                     "the given energyVector in energyConsumption.csv "
@@ -305,7 +373,7 @@ def calculate_power_demand(
     filename = os.path.join(timeseries_directory, el_demand_csv)
     shifted_elec_demand.to_csv(filename, index=False)
 
-    # save the file name of the time series and the nominal value to
+    # save the file name of the time series
     # mvs_inputs/elements/csv/energyProduction.csv
     check_inputs.add_file_name_to_energy_consumption_file(
         column=column,
@@ -508,7 +576,7 @@ def calculate_heat_demand(
     filename = os.path.join(timeseries_directory, h_demand_csv)
 
     shifted_heat_demand.to_csv(filename, index=False)
-    # save the file name of the time series and the nominal value to
+    # save the file name of the time series
     # mvs_inputs/elements/csv/energyProduction.csv
     check_inputs.add_file_name_to_energy_consumption_file(
         column=column,